refactor(ascii): centralize integration test render helpers
